Generalize Cosmos channel open init/try event
