+++ crates/cosmos/cosmos-chain-components/src/methods/event.rs
@@ -1,14 +1,9 @@
 use alloc::sync::Arc;
 
-use ibc_relayer::event::{
-    channel_open_init_try_from_abci_event, channel_open_try_try_from_abci_event,
-    extract_packet_and_write_ack_from_tx,
-};
+use ibc_relayer::event::extract_packet_and_write_ack_from_tx;
 use ibc_relayer_types::core::ics04_channel::events::{SendPacket, WriteAcknowledgement};
 use ibc_relayer_types::events::IbcEventType;
 use tendermint::abci::Event as AbciEvent;
-
-use crate::types::events::channel::{CosmosChannelOpenInitEvent, CosmosChannelOpenTryEvent};
 
 pub fn try_extract_send_packet_event(event: &Arc<AbciEvent>) -> Option<SendPacket> {
     let event_type = event.kind.parse().ok()?;
@@ -34,38 +29,6 @@ pub fn try_extract_write_ack_event(event: &Arc<AbciEvent>) -> Option<WriteAcknow
         };
 
         Some(ack)
-    } else {
-        None
-    }
-}
-
-pub fn try_extract_channel_open_init_event(
-    event: Arc<AbciEvent>,
-) -> Option<CosmosChannelOpenInitEvent> {
-    let event_type = event.kind.parse().ok()?;
-
-    if let IbcEventType::OpenInitChannel = event_type {
-        let open_init_event = channel_open_init_try_from_abci_event(&event).ok()?;
-
-        let channel_id = open_init_event.channel_id()?.clone();
-
-        Some(CosmosChannelOpenInitEvent { channel_id })
-    } else {
-        None
-    }
-}
-
-pub fn try_extract_channel_open_try_event(
-    event: Arc<AbciEvent>,
-) -> Option<CosmosChannelOpenTryEvent> {
-    let event_type = event.kind.parse().ok()?;
-
-    if let IbcEventType::OpenTryChannel = event_type {
-        let open_try_event = channel_open_try_try_from_abci_event(&event).ok()?;
-
-        let channel_id = open_try_event.channel_id()?.clone();
-
-        Some(CosmosChannelOpenTryEvent { channel_id })
     } else {
         None
     }
