+++ crates/sovereign/sovereign-relayer/src/contexts/cosmos_to_sovereign_relay.rs
@@ -11,8 +11,6 @@ use hermes_relayer_components::components::default::relay::{
 use hermes_relayer_components::relay::traits::chains::{
     CanRaiseRelayChainErrors, HasRelayChains, ProvideRelayChains,
 };
-use hermes_relayer_components::relay::traits::channel::open_handshake::CanRelayChannelOpenHandshake;
-use hermes_relayer_components::relay::traits::channel::open_try::CanRelayChannelOpenTry;
 use hermes_relayer_components::relay::traits::client_creator::CanCreateClient;
 use hermes_relayer_components::relay::traits::connection::open_handshake::CanRelayConnectionOpenHandshake;
 use hermes_relayer_components::relay::traits::connection::open_init::CanInitConnection;
